+++ frontend/src/api/bloodRequest.ts
@@ -19,7 +19,7 @@ export enum Onboardings {
 
 export type AddToPoolRequest = {
     petId: string;
-    regions: number[];
+    regions: string[];
     description?: string;
     prioritySearch: boolean;
     bloodGroupNames: string[];
@@ -98,6 +98,7 @@ export type UpdatePoolRequestResponse = {
 export type GetDonorInfoResponse = Pet & {
     taxi: boolean;
     ownerId: string;
+    regions: string[];
     ownerName: string;
     ownerPhone: string;
     availableBloodAmount: number;